requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg[binary]==3.1.18
pydantic==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.20
alembic==1.13.1
